WIP fix

+++ Xbim.ModelGeometry.Scene/IXbimTriangulatesToSimplePositionsNormalsIndices.cs
@@ -12,7 +12,7 @@ namespace Xbim.ModelGeometry.Scene
         /// </summary>
         void BeginBuild(uint PointsCount, uint TrianglesCount);
         void BeginPoints(uint PointsCount);
-        void AddPoint(float px, float py, float pz, float nx, float ny, float nz);
+        void AddPoint(double px, double py, double pz, float nx, float ny, float nz);
         void BeginTriangles(uint totalNumberTriangles);
         void AddTriangleIndex(uint index);
         void EndBuild();
